Move spanning tree counter to updaters
