+++ gerrychain/partition/partition.py
@@ -113,7 +113,6 @@ class Partition:
 
     def __getattr__(self, key):
         return self[key]
-
     def keys(self):
         return self.updaters.keys()
 
@@ -144,22 +143,6 @@ class Partition:
             {"assignment": assignment_series}, geometry=geometries
         )
         return df.plot(column="assignment", **kwargs)
-
-    def get_num_spanning_trees(self, district):
-        '''
-        Given a district number, returns the number of spanning trees in the
-        subgraph of self corresponding to the district.
-        Uses Kirchoff's theorem to compute the number of spanning trees.
-
-        :param self: :class:`gerrychain.Partition`
-        :param district: A district in self
-        :return: The number of spanning trees in the subgraph of self
-        corresponding to district
-        '''
-        graph = self.subgraphs[district]
-        laplacian = networkx.laplacian_matrix(graph)
-        L = numpy.delete(numpy.delete(laplacian.todense(), 0, 0), 1, 1)
-        return math.exp(numpy.linalg.slogdet(L)[1])
 
     @classmethod
     def from_districtr_file(cls, graph, districtr_file, updaters=None):
