+++ gerrychain/updaters/__init__.py
@@ -11,6 +11,7 @@ from .cut_edges import cut_edges, cut_edges_by_part
 from .election import Election
 from .flows import compute_edge_flows, flows_from_changes
 from .tally import DataTally, Tally
+from .spanning_trees import num_spanning_trees
 
 __all__ = [
     "flows_from_changes",
@@ -28,4 +29,5 @@ __all__ = [
     "CountySplit",
     "compute_edge_flows",
     "Election",
+    "num_spanning_trees"
 ]
